Use DI and command handler

+++ Cicada 122723/DiscordBot.cs
@@ -1,84 +1,94 @@
 ﻿using System;
 using System.IO;
 using System.Diagnostics;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
-
 using Discord;
 using Discord.WebSocket;
-
-using DSharpPlus;
-
+using Discord.Commands;
+using System.Reflection;
 using Jupiter.Commands;
 
-namespace Jupiter
+namespace Jupiter.Services
 {
-    public static class DiscordBot
+    public class DiscordBot
     {
-        public static DiscordSocketClient SocketClient;
-        public static DiscordClient Client;
-
-        private static Process LavalinkServerProcess;
-
-        public static ulong BotID { get => Client.CurrentUser.Id; }
-
         public static readonly ulong GeneralChannelID = 780678378457923595;
 
-        public static async Task InitializeAsync(string token)
+        private static Process LavalinkServerProcess;
+        private readonly IServiceProvider _serviceProvider;
+        private readonly DiscordSocketClient _client;
+        private readonly CommandService _commandService;
+
+        public DiscordBot(IServiceProvider serviceProvider,
+            DiscordSocketClient client,
+            CommandService commandService)
         {
-            //Why are we using two clients at once?
+            _serviceProvider = serviceProvider;
+            _client = client;
+            _commandService = commandService;
 
-            //Get SocketClient configuration and initialize
-            var config = new DiscordSocketConfig { MessageCacheSize = 100 };
-            SocketClient = new DiscordSocketClient(config);
+            _client.UserLeft += UserLeft;
+            _client.UserJoined += UserJoined;
 
-            //Initialize DiscordClient
-            var Client = new DiscordClient(new DiscordConfiguration()
-            {
-                Token = token,
-                TokenType = DSharpPlus.TokenType.Bot,
-            });
-
-            //Set events for Client
-            Client.GuildMemberRemoved += async (e) =>
-            {
-                await Task.Delay(100);
-                await Log(new LogMessage(LogSeverity.Info, null, e.Member.Nickname + " left"));
-            };
-
-            await Client.ConnectAsync();
-
-            //Set events for SocketClient
-            SocketClient.Log += Log;
-
-            SocketClient.UserLeft += UserLeft;
-            SocketClient.UserJoined += UserJoined;
-
-            SocketClient.Connected += Connected;
-            SocketClient.Disconnected += Disconnected;
-            SocketClient.Ready += Ready;
-
-            SocketClient.MessageReceived += MessageReceived;
-            SocketClient.MessageDeleted += MessageDeleted;
-            SocketClient.ReactionAdded += ReactionAdded;
-
-            //Login SocketClient
-            await SocketClient.LoginAsync(Discord.TokenType.Bot, token);
-            await SocketClient.StartAsync();
+            _client.Connected += Connected;
+            _client.Disconnected += Disconnected;
+            _client.Ready += Ready;
+            _client.MessageReceived += MessageReceived;
+            _client.ReactionAdded += ReactionAdded;
         }
 
-        private static async Task ReactionAdded(Cacheable<IUserMessage, ulong> cachedMessage, ISocketMessageChannel originChannel, SocketReaction reaction)
+        public async Task Start()
+        {
+            string token;
+
+            if (File.Exists(Environment.CurrentDirectory + "\\BotInfo.txt"))
+            {
+                string info = File.ReadAllText(Environment.CurrentDirectory + "\\BotInfo.txt");
+
+                token = info.Trim();
+            }
+            else
+            {
+                Console.WriteLine("BotInfo.txt file not detected. Enter bot token: ");
+                token = Console.ReadLine();
+            }
+
+            try
+            {
+                await _client.LoginAsync(TokenType.Bot, token);
+            }
+            catch (Discord.Net.HttpException ex)
+            {
+                if (ex.Reason == "401: Unauthorized")
+                {
+                    Console.WriteLine("\nToken is incorrect.");
+                    Start();
+                }
+                else
+                {
+                    Console.WriteLine("An unhandeled HttpException has occured!");
+                    Console.WriteLine(ex.Message);
+                    Console.WriteLine(ex.StackTrace);
+                    Console.WriteLine("Jupitor is quiting!");
+
+                    Environment.Exit(0);
+                }
+            }
+
+            await _client.StartAsync();
+            await _commandService.AddModulesAsync(Assembly.GetEntryAssembly(), _serviceProvider);
+        }
+
+        private async Task ReactionAdded(Cacheable<IUserMessage, ulong> cachedMessage, ISocketMessageChannel originChannel, SocketReaction reaction)
         {
             //Check, if the reaction wasn't on a help menu
-            if (reaction.UserId != BotID)
+            if (!reaction.User.Value.IsBot)
             {
                 await HelpCommand.OnReaction(reaction);
             }
             //Check if it wasn't a rock paper scissors game
             //Warning, access restriction violation happens here! Will refactor later.
-            if (RockPaperScissors.GameOver == true && reaction.UserId != BotID)
+            if (RockPaperScissors.GameOver == true && !reaction.User.Value.IsBot)
             {
                 if (reaction.MessageId.ToString().Trim() == RockPaperScissors.RpcMessageId.Trim().ToString())
                 {
@@ -86,91 +96,42 @@ namespace Jupiter
                 }
             }
         }
-        private static async Task MessageDeleted(Cacheable<IMessage, ulong> arg1, ISocketMessageChannel arg2)
+
+        private Task MessageReceived(SocketMessage msg)
         {
-            return;
-        }
-        private static async Task MessageReceived(SocketMessage msg)
-        {
-            await MessagePrinting(msg);
+            MessagePrinting(msg);
 
-            if (msg.Content.StartsWith("$jupiter say")) { await JupiterSayCommand.React(msg); return; }
+            //if (msg.Content.StartsWith("$jupiter say")) { await JupiterSayCommand.React(msg); return; }
+            //else if (msg.Content.Trim() == "$users") { await UsersCommand.React(msg); return; }
+            //else if (msg.Content.Contains("<@!782261217334001674>") || msg.Content.Contains("<@782261217334001674>")) { await TagCommand.TagBot(msg); return; }  //bot tag
+            //else if (msg.Content.StartsWith("$time")) { await TimeCommands.React(msg); return; }
+            //else if (msg.Content.Trim().StartsWith("$play")) { await MusicCommands.Play(msg, _client); }
+            //else if (msg.Content.ToLower().Trim().StartsWith("$quit")) { await MusicCommands.Quit(msg); return; }
+            //else if (msg.Content.Trim().StartsWith("$scplay")) { await MusicCommands.SoundCloudPlay(msg); return; }
+            //else if (msg.Content.ToLower().Trim().StartsWith("$stop")) { await MusicCommands.Stop(msg); return; }
+            //else if (msg.Content.ToLower().Trim().StartsWith("$resume")) { await MusicCommands.Resume(msg); return; }
+            //else if (msg.Content.ToLower().Trim().StartsWith("$pause")) { await MusicCommands.Pause(msg); return; }
+            //else if (msg.Content.StartsWith("$delete")) { await DeleteCommand.Delete(msg); return; }
+            //else if (msg.Content == "$registered_users") { await RegisteredUsersCommand.Display(msg); return; }
+            //else if (msg.Content.StartsWith("$set-tz")) { await SetTimeZoneCommand.SetTimeZone(msg); return; }
+            //else if (msg.Content.ToLower().Trim() == "$utc") { await TimeCommands.UtcTime(msg); return; }
+            //else if (msg.Content.Trim().ToLower().Contains("$warn")) { await WarnCommand.Warn(msg); return; }
+            //else if (msg.Content.Trim().ToLower() == "$help" || msg.Content.Trim().ToLower() == "$menu") { await HelpCommand.DisplayHelpMenu(msg); return; }
+            //else if (msg.Content.Trim().StartsWith("$activity")) { await ActivityCommand.Activity(msg, _client); return; 
+            //else if (msg.Content.Trim().ToLower() == "$joke") { await JokeCommand.React(msg); return; }
+            //else if (msg.Content.Trim().ToLower() == "$meme") { await MemeCommand.React(msg); return; }
+            //else if (msg.Content.Trim().Contains("$edit")) { await EditCommand.React(msg); return; }
+            //else if (msg.Content.Trim().ToLower().Contains("$rps")) { await RockPaperScissors.Start(msg); return; }
+            //else if (msg.Content.StartsWith("$8ball")) { await Ball.React(msg); return; }
+            //else if (msg.Content.Trim() == "$rr") { await RussianRoulette.React(msg); return; }
+            //else if (msg.Content.StartsWith("$rand")) { await RandCommand.React(msg); return; }
+            //else if (msg.Content.ToLower().Trim().StartsWith("$google")) { await GoogleCommand.GoogleResults(msg); return; }
+            //else if (msg.Content.StartsWith("$ttt")) { await TicTacToe.React(msg); return; }
 
-
-            else if (msg.Content.Trim() == "$users") { await UsersCommand.React(msg); return; }
-
-
-            else if (msg.Content.Contains("<@!782261217334001674>") || msg.Content.Contains("<@782261217334001674>")) { await TagCommand.TagBot(msg); return; }  //bot tag
-
-            else if (msg.Content.StartsWith("$time")) { await TimeCommands.React(msg); return; }
-
-
-            else if (msg.Content.Trim().StartsWith("$play")) { await MusicCommands.Play(msg, SocketClient); }
-
-
-            else if (msg.Content.ToLower().Trim().StartsWith("$quit")) { await MusicCommands.Quit(msg); return; }
-
-
-            else if (msg.Content.Trim().StartsWith("$scplay")) { await MusicCommands.SoundCloudPlay(msg); return; }
-
-
-            else if (msg.Content.ToLower().Trim().StartsWith("$stop")) { await MusicCommands.Stop(msg); return; }
-
-
-            else if (msg.Content.ToLower().Trim().StartsWith("$resume")) { await MusicCommands.Resume(msg); return; }
-
-
-            else if (msg.Content.ToLower().Trim().StartsWith("$pause")) { await MusicCommands.Pause(msg); return; }
-
-
-            else if (msg.Content.StartsWith("$delete")) { await DeleteCommand.Delete(msg); return; }
-
-
-            else if (msg.Content == "$registered_users") { await RegisteredUsersCommand.Display(msg); return; }
-
-
-            else if (msg.Content.StartsWith("$set-tz")) { await SetTimeZoneCommand.SetTimeZone(msg); return; }
-
-
-            else if (msg.Content.ToLower().Trim() == "$utc") { await TimeCommands.UtcTime(msg); return; }
-
-
-            else if (msg.Content.Trim().ToLower().Contains("$warn")) { await WarnCommand.Warn(msg); return; }
-
-
-            else if (msg.Content.Trim().ToLower() == "$help" || msg.Content.Trim().ToLower() == "$menu") { await HelpCommand.DisplayHelpMenu(msg); return; }
-
-
-            else if (msg.Content.Trim().StartsWith("$activity")) { await ActivityCommand.Activity(msg, SocketClient); return; }
-
-            else if (msg.Content.Trim().ToLower() == "$joke") { await JokeCommand.React(msg); return; }
-
-
-            else if (msg.Content.Trim().ToLower() == "$meme") { await MemeCommand.React(msg); return; }
-
-
-            else if (msg.Content.Trim().Contains("$edit")) { await EditCommand.React(msg); return; }
-
-
-            else if (msg.Content.Trim().ToLower().Contains("$rps")) { await RockPaperScissors.Start(msg); return; }
-
-
-            else if (msg.Content.StartsWith("$8ball")) { await Ball.React(msg); return; }
-
-
-            else if (msg.Content.Trim() == "$rr") { await RussianRoulette.React(msg); return; }
-
-
-            else if (msg.Content.StartsWith("$rand")) { await RandCommand.React(msg); return; }
-
-
-            else if (msg.Content.ToLower().Trim().StartsWith("$google")) { await GoogleCommand.GoogleResults(msg); return; }
-
-
-            else if (msg.Content.StartsWith("$ttt")) { await TicTacToe.React(msg); return; }
+            return Task.CompletedTask;
         }
 
-        static async Task MessagePrinting(SocketMessage msg)
+        private void MessagePrinting(SocketMessage msg)
         {
             Helper.ColorWriteLine(msg.Author.Username + " ", ConsoleColor.Cyan);
             Helper.ColorWriteLine("To: ", ConsoleColor.Yellow);
@@ -179,14 +140,14 @@ namespace Jupiter
             Helper.ColorWrite(msg.Content, ConsoleColor.White);
         }
 
-        private static async Task Ready()
+        private async Task Ready()
         {
             if (LavalinkServerProcess == null)
                 await Connected();
 
-            await MusicCommands.InitializeLavalink(SocketClient);
+            await MusicCommands.InitializeLavalink(_client);
         }
-        private static async Task Connected()
+        private async Task Connected()
         {
             if (File.Exists(Environment.CurrentDirectory + "\\Lavalinnk.jar"))
             {
@@ -204,52 +165,26 @@ namespace Jupiter
                 Console.WriteLine("file doesnt exists");
             }
         
-            await SocketClient.SetGameAsync("Genshin Impact", null, ActivityType.Playing);
+            await _client.SetGameAsync("Genshin Impact", null, ActivityType.Playing);
         } 
-        private static async Task Disconnected(Exception arg)
+        private async Task Disconnected(Exception arg)
         {
-            await Task.Run(() => LavalinkServerProcess.Kill());
+            if (LavalinkServerProcess != null)
+            {
+                await Task.Run(() => LavalinkServerProcess.Kill());
+            }
         }
 
-        private static async Task UserJoined(SocketGuildUser user)
+        private async Task UserJoined(SocketGuildUser user)
         {
             Console.WriteLine("NEW user joined");
             await user.Guild.GetTextChannel(GeneralChannelID).SendMessageAsync("Everyone Say Hello To " + user.Mention + ", We Hope You Will Enjoy Our Server!");
             await user.Guild.GetTextChannel(GeneralChannelID).SendMessageAsync("Hello " + user.Username + ", Nice to Meet You! My name is Jupiter, see full list of what I am capable of by typing `$help` or `$menu`");
             await user.SendMessageAsync(@$"Welcome to 'CodeVerse'. Remember to read the rules before you perform any actions in the server!");
         }
-        private static async Task UserLeft(SocketGuildUser user)
+        private async Task UserLeft(SocketGuildUser user)
         {
             await user.Guild.GetTextChannel(GeneralChannelID).SendMessageAsync(user.Username + "#" + user.Discriminator + " Has left the server...");
-        }
-
-        private static Task Log(LogMessage arg)
-        {
-            switch (arg.Severity)
-            {
-                case LogSeverity.Info:
-                    Console.ForegroundColor = ConsoleColor.Cyan;
-                    break;
-                case LogSeverity.Debug:
-                    Console.ForegroundColor = ConsoleColor.Gray;
-                    break;
-                case LogSeverity.Error:
-                    Console.ForegroundColor = ConsoleColor.Red;
-                    break;
-                case LogSeverity.Critical:
-                    Console.ForegroundColor = ConsoleColor.DarkRed;
-                    break;
-                case LogSeverity.Warning:
-                    Console.ForegroundColor = ConsoleColor.Yellow;
-                    break;
-                case LogSeverity.Verbose:
-                    Console.ForegroundColor = ConsoleColor.Gray;
-                    break;
-            }
-            Console.WriteLine(arg.ToString());
-            Console.ForegroundColor = ConsoleColor.Gray;
-
-            return Task.CompletedTask;
         }
     }
 }
